add redcut_macros crate
