+++ Cargo.toml
@@ -3,7 +3,8 @@
 members = [
     "reductstore",
     "reduct_rs",
-    "reduct_base"
+    "reduct_base",
+    "reduct_macros"
 ]
 
 resolver = "2"
