+++ reduct_macros/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "reduct-macros"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+proc-macro = true
+
+[dependencies]
+syn = "1.0.75"
+quote = "1.0.9"
